+++ src/code42cli/main.py
@@ -1,5 +1,5 @@
 import platform
-from argparse import ArgumentParser
+from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 import code42cli.securitydata.main as securitydata
 from code42cli.compat import str
@@ -7,7 +7,7 @@ from code42cli.profile import profile
 
 # If on Windows, configure console session to handle ANSI escape sequences correctly
 # source: https://bugs.python.org/issue29059
-if platform.system().lower() == "windows":
+if platform.system().lower() == u"windows":
     from ctypes import windll, c_int, byref
 
     stdout_handle = windll.kernel32.GetStdHandle(c_int(-11))
@@ -18,8 +18,17 @@ if platform.system().lower() == "windows":
 
 
 def main():
-    code42_arg_parser = ArgumentParser()
-    subcommand_parser = code42_arg_parser.add_subparsers()
+    description = u"""
+    Groups:
+        profile      - For managing Code42 settings.
+        securitydata - Tools for getting security related data, such as file events.
+    """
+    code42_arg_parser = ArgumentParser(
+        formatter_class=RawDescriptionHelpFormatter,
+        description=description,
+        usage=u"code42 <group> <subcommand> <optional args>",
+    )
+    subcommand_parser = code42_arg_parser.add_subparsers(title=u"groups")
     profile.init(subcommand_parser)
     securitydata.init_subcommand(subcommand_parser)
     _run(code42_arg_parser)
@@ -30,7 +39,7 @@ def _run(parser):
         args = parser.parse_args()
         args.func(args)
     except AttributeError as ex:
-        if str(ex) == "'Namespace' object has no attribute 'func'":
+        if str(ex) == u"'Namespace' object has no attribute 'func'":
             parser.print_help()
             return
         raise ex