+++ src/code42cli/sdk_client.py
@@ -1,16 +1,15 @@
-from py42 import debug_level
-from py42 import settings
-from py42.sdk import SDK
+import py42.sdk
+import py42.sdk.settings.debug as debug
 
 from code42cli.util import print_error
 
 
 def create_sdk(profile, is_debug_mode):
     if is_debug_mode:
-        settings.debug_level = debug_level.DEBUG
+        py42.sdk.settings.debug.level = debug.DEBUG
     try:
         password = profile.get_password()
-        return SDK.create_using_local_account(profile.authority_url, profile.username, password)
+        return py42.sdk.from_local_account(profile.authority_url, profile.username, password)
     except Exception:
         print_error(
             u"Invalid credentials or host address. "
@@ -21,7 +20,8 @@ def create_sdk(profile, is_debug_mode):
 
 def validate_connection(authority_url, username, password):
     try:
-        SDK.create_using_local_account(authority_url, username, password)
+        py42.sdk.from_local_account(authority_url, username, password)
         return True
     except:
+        print(username, password, authority_url)
         return False