+++ tests/securitydata/test_extraction.py
@@ -1,9 +1,5 @@
 import pytest
-from py42.sdk.alert_query import Actor
-from py42.sdk.file_event_query.device_query import DeviceUsername
-from py42.sdk.file_event_query.event_query import Source, EventTimestamp
-from py42.sdk.file_event_query.exposure_query import ExposureType, ProcessOwner, TabURL
-from py42.sdk.file_event_query.file_query import FilePath, FileName, SHA256, MD5
+from py42.sdk.queries.fileevents.filters import *
 
 import code42cli.securitydata.extraction as extraction_module
 from code42cli.securitydata.options import ExposureType as ExposureTypeOptions
@@ -13,7 +9,7 @@ from ..conftest import get_filter_value_from_json, get_test_date_str
 
 @pytest.fixture(autouse=True)
 def mock_42(mocker):
-    return mocker.patch("py42.sdk.SDK.create_using_local_account")
+    return mocker.patch("py42.sdk.from_local_account")
 
 
 @pytest.fixture