+++ src/code42cli/cmd/securitydata/extraction.py
@@ -18,6 +18,7 @@ from code42cli.cmd.securitydata.logger_factory import get_error_logger
 from code42cli.util import print_error, print_bold, is_interactive
 
 _EXCEPTIONS_OCCURRED = False
+_TOTAL_EVENTS = 0
 
 
 def extract(sdk, profile, output_logger, args):
@@ -140,6 +141,8 @@ def _create_event_handlers(output_logger, cursor_store):
     def handle_response(response):
         response_dict = json.loads(response.text)
         events = response_dict.get(u"fileEvents")
+        global _TOTAL_EVENTS
+        _TOTAL_EVENTS += len(events)
         for event in events:
             output_logger.info(event)
 
@@ -168,6 +171,9 @@ def _verify_compatibility_with_advanced_query(key, val):
 def _handle_result():
     if is_interactive() and _EXCEPTIONS_OCCURRED:
         print_error(u"View exceptions that occurred at [HOME]/.code42cli/log/code42_errors.")
+    global _TOTAL_EVENTS
+    if not _TOTAL_EVENTS:
+        print_to_stderr(u"No results found\n")
 
 
 def _try_append_exposure_types_filter(filters, include_non_exposure_events, exposure_types):