+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## 0.4.4 - 2020-04-01
+
+###Added
+
+- Added message to STDERR when no results are found
+
+###Fixed
+
+- Add milliseconds to end timestamp, to represent end of day with milliseconds precision.
+
+## Unreleased
+
 ## 0.4.3 - 2020-03-17
 
 ### Added